+++ TF_Settings_Web/src/Pages/Visuals/ColorPicker.tsx
@@ -1,7 +1,8 @@
 import './ColorPicker.scss';
 
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { HexAlphaColorPicker } from 'react-colorful';
+import tinycolor from 'tinycolor2';
 
 import { TabSelector } from '@/Components/Header';
 
@@ -10,13 +11,16 @@ import { CursorStyle } from './VisualsUtils';
 interface ColorPickerProps {
     cursorStyle: CursorStyle;
     updateCursorStyle: (style: CursorStyle) => void;
-    writeOutConfig: () => void;
+    writeVisualsConfigIfNew: () => void;
 }
 
 const cursorSections = ['Center Fill', 'Outer Fill', 'Center Border'] as const;
 
-const ColorPicker: React.FC<ColorPickerProps> = ({ cursorStyle, updateCursorStyle, writeOutConfig }) => {
+const ColorPicker: React.FC<ColorPickerProps> = ({ cursorStyle, updateCursorStyle, writeVisualsConfigIfNew }) => {
     const [activeTabIndex, setActiveTabIndex] = useState<number>(0);
+    const [localHex, setLocalHex] = useState<string>(cursorStyle[activeTabIndex]);
+
+    useEffect(() => setLocalHex(cursorStyle[activeTabIndex]), [cursorStyle, activeTabIndex]);
 
     return (
         <div className={'color-picker'}>
@@ -36,19 +40,26 @@ const ColorPicker: React.FC<ColorPickerProps> = ({ cursorStyle, updateCursorStyl
             <div className={'color-picker__body'}>
                 <HexAlphaColorPicker
                     color={cursorStyle[activeTabIndex]}
-                    onChange={(color) => updateCursorStyle({ ...cursorStyle, [activeTabIndex]: color })}
-                    onPointerUp={() => writeOutConfig()}
+                    onChange={(color) => {
+                        const newStyle: CursorStyle = [...cursorStyle];
+                        newStyle[activeTabIndex] = color;
+                        updateCursorStyle(newStyle);
+                    }}
                 />
                 <input
                     type="text"
                     className={'color-picker__body__text'}
-                    value={cursorStyle[activeTabIndex].toUpperCase()}
-                    onChange={(e) => updateCursorStyle({ ...cursorStyle, [activeTabIndex]: e.target.value })}
-                    onBlur={() => writeOutConfig()}
-                    onKeyDown={(e) => {
-                        if (e.key === 'Enter') {
-                            writeOutConfig();
+                    value={localHex.toUpperCase()}
+                    onChange={(e) => {
+                        const color = e.target.value;
+                        if (!tinycolor(color).isValid()) {
+                            setLocalHex(color);
+                            return;
                         }
+                        const newStyle: CursorStyle = [...cursorStyle];
+                        newStyle[activeTabIndex] = color;
+                        updateCursorStyle(newStyle);
+                        writeVisualsConfigIfNew();
                     }}
                 />
             </div>